Move constant, split error handler to functions
